worker/requirements.txt
mediapipe
opencv-python
numpy
requests
python-dotenv
